requirements.txt
google-genai
playwright
pydantic
pandas
python-dotenv